Create three squares with different colors
One of the squares orbits around the center of the screen, another one
follows the mouse position and orbits it in an interesting pattern, and
the last one is fixed in a corner. The squares should be able to change
size and/or rotation speed based on the mouse position.

This closes #24