Create a `Canvas` and `Box` component
A `Box` or 3D cube is rendered almost in the center of the screen. The
cube is blue and changes to pink when hovered over. The cube also scales
up when clicked. The `Canvas` component is used to create a 3D scene
using the `@react-three/fiber` library. The `Box` component is a
custom component that uses the `mesh` element from Three.js to create a
3D box. The `ambientLight`, `spotLight`, and `pointLight` components
are used to add lighting to the scene. The `Box` component is positioned
at `[-1.5, 0, 0]` in the 3D space.

This closes #15