Define function to create transformation matrix
This function creates a transformation matrix for scaling, rotation,
and translation in 2D space. By applying this matrix to a set of points,
you can transform those points accordingly.

This is accomplished by multiplying the translation matrix (T),
the rotation matrix (R), and the scaling matrix (S) in that order.