Display and export animated GIF of contours
This animation shows the contours of the shapes in the image. The
contours are drawn in green. The animation shows the contours being
drawn one by one, with a delay of 0.5 seconds between each contour.
With a blue text listing some properties of the shape and a red circle
around the vecinity of the centroid of the shape.